+++ fedn/cli/__init__.py
@@ -1,3 +1,4 @@
+from .client_cmd import client_cmd  # noqa: F401
 from .main import main  # noqa: F401
-from .run_cmd import run_cmd  # noqa: F401
 from .package_cmd import package_cmd  # noqa: F401
+from .run_cmd import run_cmd  # noqa: F401
